api/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
cx-Oracle==8.3.0
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.1
scikit-learn==1.0.1
gunicorn==20.1.0 